Add files via upload
Led Sequencer
LED SEQUENCER

CD4017

CD4017 has the same leg connections as 7400 and 74ls00 have the same features as in CD4017, cd series integrated datasheet files. LS indicates low power schottky. Their names start with 40 or 140. They can work with a supply between 3v…15 volts, such as 4001…14001. Power consumption is less than TTL ICs. Propagation delay times are higher than TTL ICs.

 

The 4017 IC is a 16-legged IC.

The enable, reset, carry out pins on the 4017 IC are connected to the chassis when using the circuit. After the necessary energy is given to the chassis and pover ends of the integrated and the clock pin in the circuit receives the necessary trigger signal, the integrated outputs are activated at certain intervals. If we look at the circuit, with each triggering signal coming to the 4017 IC, the Q0 output becomes active first from the 4017 outputs and the led connected to this output lights up. Then the other outputs become active respectively and the LED belonging to that output gives light. At the same time, only one output becomes active and the LED of that output lights up and this continues periodically.

 

NE555DR

 

555 IC is an IC that is used for many purposes. It is a stable integrated circuit that can be used to generate oscillation, time delay and pulse signal. The timing range can be from microseconds to hours. In addition, it has a widespread usage area with its adjustable output frequency. Timer is one of the most preferred integrated circuits among integrated circuits. This integrated low cost has stability at high frequencies. The reason for this is that the R and C elements connected to the IC can be selected at different values ​​and the timing time can be adjusted easily.

 

Pin and Descriptions

 

1) GND (Ground);

Earth Leg. Also called ground reference voltage. voltage value; is 0.

2) TRIGGER;

It is called Trigger leg in Turkish. At this input value, when the voltage value drops to half of the control voltage, a time interval starts and the Output pin becomes HIGH.

3) OUTPUT

It is called the Output leg in Turkish. It provides a voltage of approximately 1.7V below the supply voltage at high output level.

4) RESET

When this leg is LOW (Logic 0) the circuit resets. Also, this leg has no connection with any other leg.

5) CONTROL VOLTAGE

This leg is connected to 2*Vcc/3 in the internal circuit. If desired, the timing period can be changed by changing the voltage of this leg. The noise that may occur during normal operation can be prevented by connecting it to ground with a capacitor with a small capacity.

6) THRESHOLD

It is called the threshold leg in Turkish. If the voltage on this leg exceeds 2*Vcc/3, our output goes LOW (Logic 0) and the flip flop resets.

7) DISCHARGE

It is called discharge leg in Turkish. This leg is connected to the collector leg of an npn transistor in the internal circuit. It provides the charge and discharge of the capacitor according to the change of resistance at high and low levels.

8) Vcc

It is called Feed leg in Turkish. Supply voltage is applied from this leg. The supply voltage value can vary between 4.5 and 16V.

555 IC Features

• It has different features besides high stability and low cost. These features are;

• Turn-off time is less than 2 microseconds

• The maximum processing frequency is greater than 500 KHz

• Provides timing from micro-seconds to hours

• It can give high output current (200 mA)

• It has adjustable output frequency

• It requires a supply voltage between +4.5 and +18V.